pyproject.toml
[tool.poetry]
name = "moe-rag-chatbot"
version = "0.1.0"
description = "A RAG Chatbot to help with MoE research."
authors = ["Gui Osorio <gui.osorio27@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
FastAPI = "^0.70.0"
uvicorn = "^0.15.0"
ipykernel = "^6.29.5"
pandas = "^2.2.2"
python-docx = "^1.1.2"
pymupdf = "^1.24.7"
openai = "^1.35.9"
python-dotenv = "^1.0.1"
pinecone-client = "^4.1.2"
jinja2 = "^3.1.4"
python-multipart = "^0.0.9"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
